Physics changes part 1

+++ AdvancedGamesEngineering/Assets/Scripts/Moon.cs
@@ -5,11 +5,11 @@ using UnityEngine;
 public class Moon : MonoBehaviour
 {
     public int planetID; 
-    public float mass; 
+    public double mass; 
     public Vector3 scale = new Vector3(1, 1, 1); 
     public Vector3 position;
 
     public void CalculateProperties(){
-        scale *= (mass * 2);
+        scale *= ((float)mass * 2);
     }
 }
